refactor: extract personalization settings into PersonalizationCard component
