+++ src/i18n/de.json
@@ -134,7 +134,7 @@
     }
   },
   "settings": {
-    "general": "Allgemein",
+    "general": "Personalisierung",
     "language": "Sprache",
     "theme": "Thema",
     "themeLight": "Hell",
