+++ src/i18n/en.json
@@ -12,7 +12,7 @@
     "config": "Settings"
   },
   "settings": {
-    "general": "General Settings",
+    "general": "Personalization",
     "language": "Language",
     "theme": "Theme",
     "themeLight": "Light",
