+++ src/i18n/es.json
@@ -12,7 +12,7 @@
     "config": "Ajustes"
   },
   "settings": {
-    "general": "Ajustes Generales",
+    "general": "Personalización",
     "language": "Idioma",
     "theme": "Tema",
     "themeLight": "Claro",
