+++ src/i18n/fr.json
@@ -134,7 +134,7 @@
     }
   },
   "settings": {
-    "general": "Général",
+    "general": "Personnalisation",
     "language": "Langue",
     "theme": "Thème",
     "themeLight": "Clair",
