+++ src/i18n/it.json
@@ -134,7 +134,7 @@
     }
   },
   "settings": {
-    "general": "Generale",
+    "general": "Personalizzazione",
     "language": "Lingua",
     "theme": "Tema",
     "themeLight": "Chiaro",
