+++ src/i18n/pt.json
@@ -134,7 +134,7 @@
     }
   },
   "settings": {
-    "general": "Geral",
+    "general": "Personalização",
     "language": "Idioma",
     "theme": "Tema",
     "themeLight": "Claro",
